Changing file structure for pages. Converted last homepage into functional landinpage component
