--- src/HomePage/Home.tsx
@@ -1,2 +1,0 @@
-import React from 'react';
-import EventInjection from '../Components/eventinjection';
