+++ src/Pages/HomePage/HomePage.tsx
@@ -1,0 +1,37 @@
+import React from 'react';
+import "./h.css";
+import { Link } from 'react-router-dom';
+import StockGraph from '../../Components/StockGraph/stockGraph';
+import SessionSearch from '../../Components/sessionSearch';
+import ScrollableButtonList from '../../Components/SessionList/scrollableButtonList';
+import GeneralButton from '../../Components/generalButton';
+import Button from '@material-ui/core/Button';
+import CreateSessionModal from '../../Components/CreateSessionModal/CreateSessionModal'
+import MarketWindow from '../../Pages/OldPages/MarketWindow/marketwindow';
+import LogInModal from '../../Components/LogInModal/LogInModal';
+import SignUpModal from '../../Components/SignUpModal/SignUpModal';
+import Dialog from '@material-ui/core/Dialog';
+import DialogActions from '@material-ui/core/DialogActions';
+import DialogContent from '@material-ui/core/DialogContent';
+import DialogContentText from '@material-ui/core/DialogContentText';
+import DialogTitle from '@material-ui/core/DialogTitle';
+import { useHistory } from "react-router-dom";
+import { connect } from 'react-redux';
+import changeSession from '../../redux/actions';
+import changeSessionID from '../../redux/actions';
+
+type HomePageProps = {
+    history: any,
+    sessionData: any,
+    dispatch: any
+};
+
+function HomePage(props:HomePageProps){
+    
+
+    return (
+        <div>
+            
+        </div>
+    );
+}
