+++ src/Pages/OldPages/StockData/stockdata.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import firebase from 'firebase';
-import StockGraph from '../../Components/StockGraph/stockGraph';
-import BuyModal from '../../Components/BuyModal/BuyModal';
-import SellModal from '../../Components/SellModal/SellModal';
+import StockGraph from '../../../Components/StockGraph/stockGraph';
+import BuyModal from '../../../Components/BuyModal/BuyModal';
+import SellModal from '../../../Components/SellModal/SellModal';
 import { Link } from 'react-router-dom';
 
 type SDProps = {
